+++ include/pipeline.h
@@ -3,7 +3,7 @@
 #define private static
 #define SIZE 256
 
-extern const int NUM_BUILTINS;
+extern const size_t NUM_BUILTINS;
 
 int moss_execute_pipeline(char **);
 int moss_launch(char **);
