+++ src/modules/builtins/builtin.h
@@ -16,7 +16,7 @@ struct builtin
 };
 
 extern const struct builtin builtins[];
-extern const int NUM_BUILTINS;
+extern const size_t NUM_BUILTINS;
 
 int moss_cd(char **);
 int moss_help(char **);
